Jenkinsfile: Add a minimal Jenkinsfile to test-description
A minimal Jenkinsfile, which only triggers the run of the
"cirt-scheduler" Jenkins Job, has the advantage, that the state of the
different cirt-scheduler runs can be easily mapped to a dedicated
test-description (when several test-descriptions for various tests is
used).

The Jenkins Job for this Jenkinsfile has to be a multibranch project.

It could be improved by the use of the "Copy Artifact Plugin"
(https://wiki.jenkins.io/display/JENKINS/Copy+Artifact+Plugin).

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>