README: Add description how to setup a Jenkins multibranch pipeline job
A multibranch pipeline job makes it possible to have only a single git
repository describing similar but not identical tests and eases the
comparison between subsequent test run results of a single test
(described by a single branch).

For CI-RT the branches in the test-description repository are used to
to store the test description and the minimal Jenkinsfile per kernel
version.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>